+++ src/main/java/com/tightdb/lib/DateCursorColumn.java
@@ -13,12 +13,12 @@ public class DateCursorColumn<Cursor, Query> extends AbstractColumn<Date, Cursor
 
 	@Override
 	public Date get() {
-		return cursor.rowset.getDate(columnIndex, (int) cursor.getPosition());
+		return cursor.rowset.getDate(columnIndex, cursor.getPosition());
 	}
 
 	@Override
 	public void set(Date value) {
-		cursor.rowset.setDate(columnIndex, (int) cursor.getPosition(), value);
+		cursor.rowset.setDate(columnIndex, cursor.getPosition(), value);
 	}
 
 	@Override
